refactor: 유효성 검사가 인풋 값 변경때마다 동작하도록 수정
